+++ src/LcdCanvasIconDef.h
@@ -1,0 +1,25 @@
+#pragma once
+
+enum class IconIndex_t {
+    UNDEF = -1,
+    TITLE = 0,
+    ARTIST,
+    ALBUM,
+    FOLDER,
+    FILE,
+    VOLUME,
+    BATTERY,
+    YEAR,
+    GEAR,
+    CHECKED,
+    LEFTARROW,
+    _16BIT,
+    _24BIT,
+    _32BIT,
+    _44_1KHZ,
+    _48_0KHZ,
+    _88_2KHZ,
+    _96_0KHZ,
+    _176_4KHZ,
+    _192_0KHZ
+};